Extract single-pad junction solver with debug visualization
